+++ lib/api-base/fastApiContainer.ts
@@ -24,7 +24,8 @@ import { dump as yamlDump } from 'js-yaml';
 
 import { ECSCluster } from './ecsCluster';
 import { BaseProps } from '../schema';
-import { Ec2Metadata, EcsSourceType } from '../configSchema';
+import { Ec2Metadata } from '../configSchema';
+import { EcsSourceType } from '../cdk';
 import { Vpc } from '../networking/vpc';
 
 // This is the amount of memory to buffer (or subtract off) from the total instance memory, if we don't include this,
