+++ lib/user-interface/react/vite.config.ts
@@ -19,5 +19,5 @@ import { defineConfig, splitVendorChunkPlugin } from 'vite';
 
 // https://vitejs.dev/config/
 export default defineConfig({
- plugins: [react(), splitVendorChunkPlugin()],
+    plugins: [react(), splitVendorChunkPlugin()],
 });
